Login no redirecciona

+++ server/public/src/admin/pages/AdminPages/Diary/DeleteUsertoSchedule.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 
-export const DeleteUsertoDiary = () => {
-
+export const DeleteUsertoSchedule = () => {
+  const [idInput, setIdInput] = useState('');
   const [usuarioInput, setUsuarioInput] = useState('');
 
 
@@ -12,33 +12,39 @@ export const DeleteUsertoDiary = () => {
 
     event.preventDefault();
     try {
-      debugger
+
       if (usuarioInput === "" || undefined) {
-        alert(`Es obligatorio el id de usuario`)
+        alert(`Todos los campos son obligatorios`)
         return;
-      } 
+      }
+
+      if (idInput === "" || undefined) {
+        alert(`Todos los campos son obligatorios`)
+        return;
+      }
 
       let _datos = {
         usuario: usuarioInput,
         rol: "USER_ROLE"
       }
 
-      const response = await fetch(`${url}api/agendas/agenda/`, {
+      const response = await fetch(`${url}api/agendas/agenda/${idInput}`, {
         method: "DELETE",
         body: JSON.stringify(_datos),
         headers: {
           "Content-Type": "application/json",
         }
 
-      });
+      });  
+
       const data = await response.json();
   
       if (response.status !== 200) {
         if (response.status === 400) {
-          alert(`El usuario no existe en la agenda.`)
+          alert(`El usuario no existe en este horario.`)
           return;
         } else {          
-          throw data.error || new Error(`Request failed with status ${response.status}`);
+          throw data.error || new Error(`Request failed with status ${response.message}`);
         }
       }      
       alert(`Usuario ${usuarioInput} borrado`)
@@ -46,7 +52,7 @@ export const DeleteUsertoDiary = () => {
     } catch (error) {
       // Consider implementing your own error handling logic here
       console.error(error);
-      alert(error.message);
+      alert("No se ha podido borrar al usuario del horario.");
     }
   }
 
@@ -56,8 +62,17 @@ export const DeleteUsertoDiary = () => {
     <>
       <br />
       <form className='formAdmin' onSubmit={onSubmitDeleteUseronSchedule} >
-        <h1>Dar de baja al usuario en la agenda</h1>
-        <br/>
+        <h1>Borrar usuario del horario</h1>
+        <br />
+        <label>Id horario </label>
+        <input
+          type="text"
+          name="id"
+          placeholder="Id"
+          value={idInput}
+          onChange={(e) => setIdInput(e.target.value)}
+        />
+        <br />
         <label>Id usuario </label>
         <input
           type="text"
@@ -66,11 +81,10 @@ export const DeleteUsertoDiary = () => {
           value={usuarioInput}
           onChange={(e) => setUsuarioInput(e.target.value)}
         />
-        <input className='botonForm' type="submit" value="Dar de baja usuario" />
+        <input className='botonForm' type="submit" value="Borrar usuario" />
       </form>
 
 
     </>
   )
- 
 }
