+++ server/public/src/admin/pages/AdminPages/User/DeleteUser.jsx
@@ -1,17 +1,23 @@
 import { useState } from 'react';
 
-export const DeleteActivity = () => {
-  const [actividadInput, setActividadInput] = useState('');
+export const DeleteUser = () => {
+
+  const [nameInput, setNameInput] = useState('');
   const [idInput, setIdInput] = useState('');
   const url = 'http://localhost:8080/'
 
   //Función borrar usuario
-  async function onSubmitDeleteActivity(event) {
+  async function onSubmitDeleteUser(event) {
 
     event.preventDefault();
+    if(!idInput){
+      alert("Es obligatorio introducir el id del usuario")
+      return;
+    }
+
     try {
 
-      const response = await fetch(`${url}api/actividades/${idInput}`, {
+      const response = await fetch(`${url}api/usuarios/${idInput}`, {
         method: "DELETE",
         headers: {
           "Content-Type": "application/json",
@@ -20,24 +26,24 @@ export const DeleteActivity = () => {
       });
 
       const data = await response.json();
-  
+
       if (response.status !== 200) {
-        throw data.error || new Error(`Request failed with status ${response.status}`);
+        throw data.error || new Error(`${response.message}`);
       }
-      alert(`Actividad ${actividadInput} borrado`)
+      alert(`Usuario ${nameInput} borrado`)
 
     } catch (error) {
       console.error(error);
-      alert(error.message);
+      alert("No se ha podido borrar el usuario.");
     }
   }
   return (
 <>
       <br />
-      <form className='formAdmin' onSubmit={onSubmitDeleteActivity} >
-        <h1>Borrar Actividad</h1>
+      <form className='formAdmin' onSubmit={onSubmitDeleteUser} >
+        <h1>Borrar Usuario</h1>
         <br/>
-        <label>Id actividad </label>
+        <label>Id usuario </label>
         <input
           type="text"
           name="id"
@@ -46,7 +52,7 @@ export const DeleteActivity = () => {
           onChange={(e) => setIdInput(e.target.value)}
         />
         <br />
-        <input className='botonForm' type="submit" value="Borrar actividad" />
+        <input className='botonForm' type="submit" value="Borrar usuario" />
 
 
       </form>
